Refactor Main component and couple start evaluation creation hook with ui
