+++ src/features/evaluations/application/repository/evaluations-repository.ts
@@ -1,0 +1,5 @@
+import useStartEvaluationCreationUsecase from "../usecases/useStartEvaluationCreationUsecase";
+
+export const evaluationsRepository = {
+  useStartEvaluationCreation: () => useStartEvaluationCreationUsecase(),
+};
